+++ tests/runner.js
@@ -14,6 +14,7 @@ function addFiles(mocha, files) {
   glob.sync(root + files).forEach(mocha.addFile.bind(mocha));
 }
 
+addFiles(mocha, 'jshint.spec.js');
 addFiles(mocha, '/**/*-nodetest.js');
 
 if (arg === 'all') {