+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [0.0.2] - 2015-11-22
+- Change default `manifestRoot` to ```/revisions/${revisionKey}```.
+- Add jshint tests and fix style errors.
+- Deleted unused ember-cli default files.
+
 ## [0.0.1] - 2015-10-13
 - Rework api, added funtionality to create manifest file.